// core/schemas/entities.ts
export type DesignChallengeStatus = 'Open' | 'Voting' | 'Complete';
// ...
export interface DesignChallengeEntity {
    id: string;
    title: string;
    description: string;
    reward: number; // in ARCHI
    status: DesignChallengeStatus;
    endsAt: string;
    winnerId?: string;
}